requirements.txt
nose
sphinx
numpy
libKMCUDA
http://download.pytorch.org/whl/cu80/torch-0.3.1-cp35-cp35m-linux_x86_64.whl
torchvision
